Refactor pulling runner logs
